Split commands.go into domain sub-packages
Extract pure parsing and formatting logic from the monolithic
commands.go (~2500 lines) into focused sub-packages under
internal/server/commands/:

- layout/     — split/spawn arg parsing, direction naming
- layout/tree — move arg parsing
- input/keys  — key parsing, special keys map, hex encoding
- listing/    — pane/window/client list formatting
- meta/       — collection metadata parsing (PR/issue add/rm)
- remote/     — reload exec path parsing
- wait/       — wait-* flag parsing, busy helpers

Command handler functions remain in the server package (they need
CommandContext), split into domain-specific files (commands_layout.go,
commands_input.go, etc.). Thin wrapper functions delegate to the
sub-packages, preserving the existing internal API for tests.
